The Lazy Shutdown - ver 1.0
This is the Lazy Shutdown button, made by me! Also, this is my first project on GitHub EVER!!! If you are too lazy to click on your Start bar, click power button, click "shutdownn" button and, this is too hard! So, with this .bat file you can click the app on your desktop, press any key, click close (this is only for this version, I will try and fix it in the last version; ver 2.0) or just wait and enjoy!
©️Copyright, TheDotBatNewbie, 2023-2023